requirements.txt
qq-botpy==1.2.1
peewee==3.17.6
spark_ai_python==0.4.5